Better layout management

+++ app/src/main/java/com/imam/cineme/model/Movie.java
@@ -173,7 +173,7 @@ public class Movie implements Parcelable {
         return rate;
     }
 
-    public void setRate(float rate) {
+    public void setRate(double rate) {
         this.rate = rate;
     }
